Change type of data in got_introspection to gpointer.
On some machines (particularly inside scratchbox) this field would
break GCC's strict aliasing rules with -O2 or higher by casting
from a SubscriptionData to a gpointer.  Since we don't strictly
need it typed as a SubscriptionData and we can let GCC to strict
aliasing with it as a gpointer we can safely type it as just a
gpointer and prevent warnings/errors.